+++ src/main/java/gregtech/api/graphs/consumers/NodeGTBaseMetaTile.java
@@ -6,13 +6,13 @@ import java.util.ArrayList;
 
 //consumer for gt machines
 public class NodeGTBaseMetaTile extends ConsumerNode {
-    public NodeGTBaseMetaTile(int aNodeValue, BaseMetaTileEntity aTileEntity, int aSide, ArrayList<ConsumerNode> aConsumers) {
+    public NodeGTBaseMetaTile(int aNodeValue, BaseMetaTileEntity aTileEntity, byte aSide, ArrayList<ConsumerNode> aConsumers) {
         super(aNodeValue, aTileEntity, aSide, aConsumers);
     }
 
     @Override
     public int injectEnergy(int aVoltage, int aMaxApms) {
-        return (int)((IEnergyConnected) mTileEntity).injectEnergyUnits((byte) mSide,aVoltage, aMaxApms);
+        return (int)((IEnergyConnected) mTileEntity).injectEnergyUnits(mSide,aVoltage, aMaxApms);
     }
 
     @Override
